% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_class.R
\name{BikeClient}
\alias{BikeClient}
\title{R6 class Representing a connection to the bike data client.}
\description{
R6 class Representing a connection to the bike data client.

R6 class Representing a connection to the bike data client.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{stations}}{df of stations}

\item{\code{status}}{df of statuses}

\item{\code{last_updated}}{character, when last updated}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BikeClient$new()}}
\item \href{#method-update_status}{\code{BikeClient$update_status()}}
\item \href{#method-clone}{\code{BikeClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new BikeClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BikeClient$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_status"></a>}}
\subsection{Method \code{update_status()}}{
Update current statuses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BikeClient$update_status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BikeClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
